refactor: Move feature-specific managers and configs from core to features
