+++ src/core/events/playerSpawn.ts
@@ -3,7 +3,7 @@ import * as mc from '@minecraft/server';
 import { getConfig } from '@core/configManager.js';
 import { getKitsConfig } from '@core/configurations.js';
 import { frozenTag, vanishedTag } from '@core/constants.js';
-import { getKit, giveKitItems } from '@core/kitsManager.js';
+import { getKit, giveKitItems } from '@features/kits/kitsManager.js';
 import { debugLog, infoLog } from '@core/logger.js';
 import { sendMessage } from '@core/messaging.js';
 import { getOrCreatePlayer, updatePlayerData } from '@core/playerDataManager.js';
