+++ src/core/mobDeathEvents.ts
@@ -10,7 +10,7 @@ import * as lastHitManager from '@core/lastHitManager.js';
 import { infoLog } from '@core/logger.js';
 import { getPlayerFromCache } from '@core/playerCache.js';
 import { getPlayer, incrementDeathCount, incrementKillCount, incrementKillStreak, incrementPlayerBalance, resetKillStreak } from '@core/playerDataManager.js';
-import { handlePvPDeath } from '@core/pvpManager.js';
+import { handlePvPDeath } from '@features/essentials/pvpManager.js';
 import { formatCurrency } from '@core/utils.js';
 import { saveLastLocation } from '@features/teleportation/teleportUtils.js';
