+++ src/core/ui/panels/infoPanel.ts
@@ -3,9 +3,9 @@ import * as mc from '@minecraft/server';
 import { ActionFormResponse, ModalFormData, ModalFormResponse } from '@minecraft/server-ui';
 
 import { getConfig } from '@core/configManager.js';
-import * as helpfulLinksManager from '@core/helpfulLinksManager.js';
+import * as helpfulLinksManager from '@features/essentials/helpfulLinksManager.js';
 import { getOrCreatePlayer, type PlayerData } from '@core/playerDataManager.js';
-import * as rulesManager from '@core/rulesManager.js';
+import * as rulesManager from '@features/essentials/rulesManager.js';
 import { showPanel } from '@core/uiManager.js';
 import { isDefined, isNonEmptyString } from '@lib/guards.js';
 import { getStaticMenuItems } from '@ui/panelBuilder.js';
