+++ src/core/ui/panels/sidebarPanel.ts
@@ -2,7 +2,7 @@ import * as mc from '@minecraft/server';
 import { ActionFormResponse, ModalFormData, ModalFormResponse } from '@minecraft/server-ui';
 
 import { getSidebarConfig, saveSidebarConfig } from '@core/configurations.js';
-import { forceUpdate } from '@core/sidebarManager.js';
+import { forceUpdate } from '@features/sidebar/sidebarManager.js';
 import { showPanel } from '@core/uiManager.js';
 import { isDefined, isNonEmptyString } from '@lib/guards.js';
 import { getStaticMenuItems } from '@ui/panelBuilder.js';
