+++ src/core/ui/uiUtils.ts
@@ -18,9 +18,9 @@ import {
 } from '@core/configurations.js';
 import type { PlayerData } from '@core/playerDataManager.js';
 import ranksConfig from '@core/ranksConfig.default.js';
-import { spawnConfig } from '@core/spawnConfig.default.js';
+import { spawnConfig } from '@features/essentials/spawnConfig.default.js';
 import { showPanel } from '@core/uiManager.js';
-import { xrayConfig } from '@core/xrayConfig.default.js';
+import { xrayConfig } from '@features/anticheat/xrayConfig.default.js';
 import { AnticheatConfig, getAnticheatConfig, saveAnticheatConfig } from '@features/anticheat/anticheatConfigLoader.js';
 import { auctionHouseConfig } from '@features/auctionHouse/auctionHouseConfig.default.js';
 import { economyConfig } from '@features/economy/economyConfig.js';
