+++ src/features/economy/commands/balance.ts
@@ -2,7 +2,7 @@ import * as mc from '@minecraft/server';
 
 import { CommandExecutor, CustomCommand } from '@commands/commandManager.js';
 import { getConfig } from '@core/configManager.js';
-import { getLeaderboard } from '@core/leaderboardManager.js';
+import { getLeaderboard } from '@features/economy/leaderboardManager.js';
 import { sendMessage } from '@core/messaging.js';
 import { getOrCreatePlayer, getPlayerIdByName, getPlayerNameById, loadPlayerData } from '@core/playerDataManager.js';
 import { formatCurrency, resolveTarget } from '@core/utils.js';
