+++ src/features/economy/commands/bounty.ts
@@ -1,6 +1,6 @@
 import * as mc from '@minecraft/server';
 
-import * as bountyManager from '@core/bountyManager.js';
+import * as bountyManager from '@features/economy/bountyManager.js';
 import { getConfig } from '@core/configManager.js';
 import { sendMessage } from '@core/messaging.js';
 import { getOrCreatePlayer, getPlayerIdByName, getPlayerNameById, incrementPlayerBalance } from '@core/playerDataManager.js';
