+++ src/features/essentials/commands/floatingtext.ts
@@ -1,7 +1,7 @@
 import * as mc from '@minecraft/server';
 
 import { CommandExecutor, CustomCommand } from '@commands/commandManager.js';
-import * as floatingTextManager from '@core/floatingTextManager.js';
+import * as floatingTextManager from '@features/essentials/floatingTextManager.js';
 import { isNonEmptyString } from '@lib/guards.js';
 
 const command: CustomCommand = {
