+++ src/features/essentials/helpfulLinksManager.ts
