+++ src/features/essentials/restartManager.ts
