+++ src/features/essentials/rulesManager.ts
